Move the dialogs used in the editor to the body
The jQuery form validate plugin throws an exception when the editor is in a modal window and another modal is displayed through the editor.  The reason for this seems to be the fact that the editor dialogs contain forms.  By moving the editor dialogs to the body, this is no longer an issue.