Moved dialogs to the body
When the editor was placed in a bootstrap modal and jquery validate was used, errors were thrown when a modal was displayed, image, video, etc... form the editor.  By moving the dialogs to the body this error was not thrown any longer because there were no longer nested forms within the modal.