Remove existing class on format block
This fixes #3159.

If you have added styles to the styles menu using the styleTags option
that use css classes and you have more than one style that uses the same
tag (e.g. two styles that both use <p>) then when changing
the style of a block from one to another, the class names of both styles
are added to the block instead of the class from the new style replacing
the class from the old one.

This fixes the issue by removing any classes from a block in
formatBlock.

To reproduce:

- Add the following styles to the page:
  <style type="text/css">
    .red { color: red; }
    .blue { color: blue; }
  </style>

- Add the following styleTags to summernote:

     $('.summernote').summernote({
        styleTags: [
          'p',
          {title: 'Red', tag: 'p', className: 'red', value: 'p'},
          {title: 'Blue', tag: 'p', className: 'blue', value: 'p'},
        ]
      });

- Set style to "Red" and type a paragraph.
- Change style to "Blue".
-Switch to code view - note that the class of the block has both class
names i.e. <p class="red blue">.

Additionaly if you now set the style back to normal the block is still
blue since the class="blue" is not removed.