Remove Husky from dependencies (#4699)
This PR removes Husky from the Summernote package.json. With the
introduction of squash merge in the Summernote repository, the need
for Husky to prevent unnecessary commits has significantly decreased.

Additionally, there is an issue with Husky causing errors during
installation when using npm with Node version 20.12.0. The problem
arises from the Husky command in the preinstall script of the
package.json, leading to complications for users.

By removing Husky from the dependencies, we aim to enhance the
installation experience and reduce potential issues.